Email_Validation_Tool/email_vaildator_project/requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosmtplib==4.0.2
asgiref==3.9.2
attrs==25.3.0
cffi==2.0.0
Django==5.2.6
djangorestframework==3.16.1
dnspython==2.8.0
email-validator==2.3.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
nest-asyncio==1.6.0
propcache==0.3.2
pycares==4.11.0
pycparser==2.23
sqlparse==0.5.3
yarl==1.20.1
